requirements.txt
streamlit>=1.18.0
onnxruntime>=1.15.1
numpy>=1.24.0
opencv-python>=4.7.0
Pillow>=9.4.0